Extract wollo slider
