+++ src/app/components/progress/index.js
@@ -1,5 +1,6 @@
 import React, {Component} from 'react';
 import {View, Text, Image} from 'react-native';
+import RNModal from 'react-native-modal';
 import styles from './styles';
 import Bar from './bar';
 import Button from '../button';
@@ -45,28 +46,30 @@ export default class Progress extends Component {
         const errorMessage = (error && error.message) || error;
 
         return (
-            <View style={styles.overlay}>
-                <View style={styles.container}>
-                    <Text style={styles.title}>{title}</Text>
-                    <Paragraph small style={errorMessage ? [styles.text, styles.error] : styles.text}>
-                        {errorMessage || text}
-                    </Paragraph>
-                    {complete ? (
-                        <Image style={styles.check} source={require('./images/check.png')}/>
-                    ) : (
-                        <Bar active={active} error={error}/>
-                    )}
-                    <View style={styles.inner}>
-                        {buttonLabel &&
-                            <Button
-                                theme="outline"
-                                disabled={!error && !complete}
-                                label={buttonLabel}
-                                onPress={onPress}
-                            />}
+            <RNModal isVisible={true} style={{margin: 0}}>
+                <View style={styles.overlay}>
+                    <View style={styles.container}>
+                        <Text style={styles.title}>{title}</Text>
+                        <Paragraph small style={errorMessage ? [styles.text, styles.error] : styles.text}>
+                            {errorMessage || text}
+                        </Paragraph>
+                        {complete ? (
+                            <Image style={styles.check} source={require('./images/check.png')}/>
+                        ) : (
+                            <Bar active={active} error={error}/>
+                        )}
+                        <View style={styles.inner}>
+                            {buttonLabel &&
+                                <Button
+                                    theme="outline"
+                                    disabled={!error && !complete}
+                                    label={buttonLabel}
+                                    onPress={onPress}
+                                />}
+                        </View>
                     </View>
                 </View>
-            </View>
+            </RNModal>
         );
     }
 }
