+++ src/app/screens/child-dash/index.js
@@ -9,6 +9,7 @@ import StepModule from '../../components/step-module';
 import KidAvatar from '../../components/kid-avatar';
 import ActionPanel from '../../components/action-panel';
 import ActionSheet from '../../components/action-sheet';
+import WolloSendSlider from 'app/components/wollo-send-slider';
 import styles from './styles';
 import {familyDeleteAllowance, familyDeleteTask} from '../../actions';
 
@@ -96,6 +97,7 @@ export class ChildDash extends Component {
             exchange,
             error,
             baseCurrency,
+            balance,
         } = this.props;
 
         const loading = !exchange && !error;
@@ -179,6 +181,19 @@ export class ChildDash extends Component {
                                 boxButton
                             >
                                 <View style={styles.box} />
+                            </ActionPanel>
+
+                            <ActionPanel
+                                title="Gift"
+                                style={styles.panel}
+                                containerOnly={true}
+                            >
+                                <View style={styles.box}>
+                                    <WolloSendSlider
+                                        name={kid.name}
+                                        address={kid.address}
+                                    />
+                                </View>
                             </ActionPanel>
                         </View>
                     )}
